apps/backend-rag/requirements.txt
# ZANTARA RAG Backend Dependencies v5.4 - DEPENDENCY CONFLICTS RESOLVED
# Last updated: 2025-12-03

# Core Framework
fastapi>=0.115.0  # Security fixes
uvicorn[standard]==0.24.0
sentry-sdk[fastapi]>=2.0.0  # Error monitoring
pydantic==2.9.2  # Compatible with MCP (requires >=2.8.0) - 2.10 has breaking changes
pydantic-settings==2.5.2  # Compatible with pydantic 2.9
python-multipart==0.0.6

# Vector Database
qdrant-client>=1.12.0  # Upgraded to support urllib3>=2.0.0

# Google Cloud Integration (Core for v5.3 Hybrid) - Use compatible versions
google-generativeai>=0.5.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-cloud-storage>=2.10.0
# google-cloud-aiplatform removed due to protobuf conflicts

# AI/ML - PINNED VERSIONS TO AVOID CONFLICTS
openai==1.55.0
anthropic==0.7.8
# PyTorch - Explicitly specify compatible version to avoid RuntimeError with sentence-transformers
# PyTorch 2.2.0+ required for transformers compatibility (register_pytree_node)
torch>=2.2.0,<2.3.0  # Compatible with sentence-transformers 2.7.0 and transformers
sentence-transformers==2.7.0
tiktoken>=0.5.0  # For accurate token estimation (optional, has fallback)

# Document Processing
openpyxl==3.1.2
email-validator==2.1.0
pypdf==3.17.1
PyPDF2==3.0.1
EbookLib==0.17.1
beautifulsoup4==4.12.2

# LangChain Ecosystem
langgraph>=0.0.1
langchain-core>=0.1.0
# langchain-google-genai removed to resolve conflicts

# HTTP/Network - UPGRADED TO RESOLVE CONFLICTS
httpx==0.27.2  # Pinned to 0.27.2 to avoid openai proxies conflict
requests==2.31.0
aiohttp>=3.10.0  # Required by litellm, runpod
urllib3>=2.0.0  # Required by selenium, types-requests
sse-starlette>=1.8.0  # For Server-Sent Events (Streaming)

# Web Scraping (Peraturan Spider & Putusan Spider)
playwright==1.40.0
fake-useragent==1.4.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Authentication
python-jose[cryptography]>=3.4.0  # CVE-2024-33663, CVE-2024-33664 fixed in 3.4.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
sqlmodel==0.0.14
sqlalchemy>=2.0.0

# Utilities - UPGRADED VERSIONS
python-dotenv==1.0.0
# loguru removed - using standard logging instead
pytz>=2024.2  # Required by dateparser
packaging>=23.2,<25  # Compatible with langchain-core if needed

# Development & Quality Tools
pre-commit==3.6.0
ruff==0.1.8
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
testcontainers[compose]>=4.0.0  # For integration tests with Docker containers

# Observability
prometheus-fastapi-instrumentator>=6.1.0
# OpenTelemetry removed due to protobuf conflicts with google packages
# opentelemetry-api>=1.34.1
# opentelemetry-sdk>=1.34.1
# opentelemetry-instrumentation-fastapi>=0.45b0
# opentelemetry-exporter-otlp>=1.34.1

# Data Processing
datasets>=2.19.0
pandas>=2.0.0
numpy<2.0.0  # Required for torch/sentence-transformers compatibility
psutil>=5.9.0  # Required by metrics module/tests

# PDF Processing - UPGRADED
pdfplumber>=0.11.0
pdfminer.six>=20251107  # Match pdfplumber requirement

# Protocol Buffers - Let pip resolve version
# protobuf>=5.29.0  # Removed: conflicts with google-cloud-aiplatform

# XML/HTML Processing - UPGRADED
lxml>=5.3.0  # Required by trafilatura, htmldate

# Agentic RAG Dependencies
pymupdf>=1.23.0  # For PDF rendering in Vision Service
scikit-learn>=1.3.0  # For cosine similarity in Golden Router

# Notification Services
sendgrid>=6.11.0  # For email notifications via SendGrid API
twilio>=9.0.0  # For WhatsApp and SMS notifications via Twilio API
tenacity>=8.2.3  # For retry logic with exponential backoff

# MCP (Model Context Protocol)
mcp==1.3.0  # Pinned for stability - requires pydantic>=2.8.0,<2.10
